# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toolshed Hire API",
    "description": "Rental system for Toolshed Hire. Availability is the allocation of specific tagged assets over half open date periods, enforced by a PostgreSQL GiST exclusion constraint.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/me/rentals": {
      "get": {
        "tags": [
          "hire"
        ],
        "summary": "List the signed in customer's own rentals, newest first",
        "description": "Return one page of the caller's own rentals, with no asset tag on any item.",
        "operationId": "list_my_rentals_api_me_rentals_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "The page, counted from 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "The page, counted from 1."
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "How many rentals a page holds.",
              "default": 20,
              "title": "Pagesize"
            },
            "description": "How many rentals a page holds."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rentals": {
      "get": {
        "tags": [
          "hire"
        ],
        "summary": "List rentals at any branch, the overdue first",
        "description": "Return one page of rentals, the most overdue first and then the newest.\n\nRaises:\n    ValidationFailure: If the branch code is not the code of a trading\n        branch. HTTP 422, naming `branchCode`.",
        "operationId": "list_rentals_api_rentals_get",
        "parameters": [
          {
            "name": "branchCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A branch code. Only rentals that went out there.",
              "title": "Branchcode"
            },
            "description": "A branch code. Only rentals that went out there."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RentalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only rentals in this status.",
              "title": "Status"
            },
            "description": "Only rentals in this status."
          },
          {
            "name": "overdueOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only rentals with a unit out past its due date.",
              "default": false,
              "title": "Overdueonly"
            },
            "description": "Only rentals with a unit out past its due date."
          },
          {
            "name": "customerProfileId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this customer's.",
              "title": "Customerprofileid"
            },
            "description": "Only this customer's."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "The page, counted from 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "The page, counted from 1."
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "How many rentals a page holds.",
              "default": 20,
              "title": "Pagesize"
            },
            "description": "How many rentals a page holds."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "A query parameter was refused. `errors.fields` names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{id}": {
      "get": {
        "tags": [
          "hire"
        ],
        "summary": "Return one rental, by its key or its reference",
        "description": "Return the rental, or 404 when it does not exist.\n\nRaises:\n    NotFound: If there is no such rental. HTTP 404.",
        "operationId": "read_rental_api_rentals__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "The key of the rental, or its reference.",
              "title": "Id"
            },
            "description": "The key of the rental, or its reference."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalResponse"
                }
              }
            }
          },
          "404": {
            "description": "There is no such rental.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{id}/returns": {
      "post": {
        "tags": [
          "hire"
        ],
        "summary": "Take units of a rental back, and settle the deposit when the last is back",
        "description": "Record the return of the named units in one transaction.\n\nRaises:\n    NotFound: If there is no such rental. HTTP 404.\n    BranchScopeError: If counter staff take units back at another branch. HTTP 403.\n    StateTransitionError: If a unit is already back. HTTP 409.\n    ValidationFailure: If a unit is listed twice, is not on the rental or\n        has a meter reading below the one it went out with. HTTP 422,\n        naming the field.",
        "operationId": "post_return_api_rentals__id__returns_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "The key of the rental, or its reference.",
              "title": "Id"
            },
            "description": "The key of the rental, or its reference."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller may not do this. The problem type says why, which is the role, the branch, an account on hold or an email address that is not verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "There is no such rental.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "A unit named is already back, so nothing was returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "A field was refused. `errors.fields` names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{id}/items/{itemId}/loss": {
      "post": {
        "tags": [
          "hire"
        ],
        "summary": "Record a unit more than fourteen days late as lost",
        "description": "Record the loss of one unit in one transaction.\n\nRaises:\n    NotFound: If there is no such rental, or the unit is not on it. HTTP 404.\n    BranchScopeError: If counter staff record a loss at another branch. HTTP 403.\n    StateTransitionError: If the unit is already back or not yet late\n        enough. HTTP 409.",
        "operationId": "post_loss_api_rentals__id__items__itemId__loss_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "The key of the rental, or its reference.",
              "title": "Id"
            },
            "description": "The key of the rental, or its reference."
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "The key of the rental item.",
              "title": "Itemid"
            },
            "description": "The key of the rental item."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller may not do this. The problem type says why, which is the role, the branch, an account on hold or an email address that is not verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "There is no such rental, or the unit is not on it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "The unit is already back, or it is not yet more than fourteen days past its due date, and `detail` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{id}/balance-payment": {
      "post": {
        "tags": [
          "hire"
        ],
        "summary": "Record the simulated payment of a rental's balance",
        "description": "Record the payment and settle the rental in one transaction.\n\nRaises:\n    NotFound: If there is no such rental. HTTP 404.\n    BranchScopeError: If counter staff record a payment at another branch. HTTP 403.\n    StateTransitionError: If nothing is owed. HTTP 409.",
        "operationId": "post_balance_payment_api_rentals__id__balance_payment_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "The key of the rental, or its reference.",
              "title": "Id"
            },
            "description": "The key of the rental, or its reference."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalancePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller may not do this. The problem type says why, which is the role, the branch, an account on hold or an email address that is not verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "There is no such rental.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Nothing is owed on the rental, so there is no balance to pay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "A field was refused. `errors.fields` names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalancePaymentRequest": {
        "properties": {
          "paymentReference": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Paymentreference"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paymentReference"
        ],
        "title": "BalancePaymentRequest",
        "description": "The simulated payment of a balance, by the reference the customer was given."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "errors": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "An RFC 9457 problem document.\n\n`type` is a stable slug, not a resolvable URL, because this API is not a\npublic specification and inventing a documentation URL that does not exist\nis worse than a slug that does.\n\n`requestId` is an extension member. It is the same value as the\n`X-Request-ID` response header and the `request_id` on every log record of\nthe request, so a caller who quotes it has named the exact log lines."
      },
...
      "RentalPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RentalResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "title": "RentalPageResponse",
        "description": "One page of rentals."
      },
      "RentalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "status": {
            "$ref": "#/components/schemas/RentalStatus"
          },
          "reservationId": {
            "type": "string",
            "format": "uuid",
            "title": "Reservationid"
          },
          "reservationReference": {
            "type": "string",
            "title": "Reservationreference"
          },
          "branchCode": {
            "type": "string",
            "title": "Branchcode"
          },
          "branchName": {
            "type": "string",
            "title": "Branchname"
          },
          "customerProfileId": {
            "type": "string",
            "format": "uuid",
            "title": "Customerprofileid"
          },
          "customerName": {
            "type": "string",
            "title": "Customername"
          },
          "customerPhone": {
            "type": "string",
            "title": "Customerphone"
          },
          "from": {
            "type": "string",
            "format": "date",
            "title": "From"
          },
          "dueBackOn": {
            "type": "string",
            "format": "date",
            "title": "Duebackon"
          },
          "checkedOutAt": {
            "type": "string",
            "title": "Checkedoutat"
          },
          "returnedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returnedat"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RentalItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "charges": {
            "items": {
              "$ref": "#/components/schemas/ChargeResponse"
            },
            "type": "array",
            "title": "Charges"
          },
          "depositHeld": {
            "type": "string",
            "title": "Depositheld"
          },
          "depositWithheld": {
            "type": "string",
            "title": "Depositwithheld"
          },
          "depositRefunded": {
            "type": "string",
            "title": "Depositrefunded"
          },
          "balanceDue": {
            "type": "string",
            "title": "Balancedue"
          },
          "settledAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settledat"
          },
          "agreementSigned": {
            "type": "boolean",
            "title": "Agreementsigned"
          },
          "canReturn": {
            "type": "boolean",
            "title": "Canreturn"
          },
          "settlementWaitingOn": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SettlementWait"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "reference",
          "status",
          "reservationId",
          "reservationReference",
          "branchCode",
          "branchName",
          "customerProfileId",
          "customerName",
          "customerPhone",
          "from",
          "dueBackOn",
          "checkedOutAt",
          "returnedAt",
          "items",
          "charges",
          "depositHeld",
          "depositWithheld",
          "depositRefunded",
          "balanceDue",
          "settledAt",
          "agreementSigned",
          "canReturn",
          "settlementWaitingOn"
        ],
        "title": "RentalResponse",
        "description": "One rental, as the caller is allowed to see it."
      },
      "RentalStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "OVERDUE",
          "PARTIALLY_RETURNED",
          "RETURNED",
          "SETTLED"
        ],
        "title": "RentalStatus",
        "description": "Rental lifecycle status, governed by BR-29, BR-52 and BR-53."
      },
...
      "ReturnRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReturnItemRequest"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ReturnRequest",
        "description": "One or more units of a rental that came back."
      },
      "SettlementWait": {
        "type": "string",
        "enum": [
          "ITEMS_OUT",
          "DAMAGE_ASSESSMENT",
          "BALANCE_PAYMENT"
        ],
        "title": "SettlementWait",
        "description": "What a rental is waiting on before its deposit can be settled."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}